Refactor: split functionality into logic, ui, and models modules
