+++ src/ui.rs
@@ -1,0 +1,28 @@
+/*!
+git-brust - Rust CLI tool to visualize git branch flows
+Copyright (C) 2025  Juan Luis Leal Contreras (Kuenlun)
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+
+use log::info;
+
+use crate::models::Relation;
+
+pub fn render(relations: Vec<Relation>) -> () {
+    info!("Relations:");
+    for relation in &relations {
+        info!("{}", relation);
+    }
+}
